perubahan baru

+++ app/Http/Controllers/StokController.php
@@ -4,12 +4,14 @@ namespace App\Http\Controllers;
 
 use App\Models\Stok;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth;
 
 class StokController extends Controller
 {
     public function index()
     {
-        $stok = Stok::all(); // Mengambil semua data stok
+        //agar setiap user beda beda data 
+        $stok = Stok::where('user_id', Auth::id())->get();
         return view('stock.stok', compact('stok'));
     }
 
@@ -26,14 +28,21 @@ class StokController extends Controller
             'ram' => 'required',
             'stok' => 'required|integer',
         ]);
-
-        Stok::create($request->all()); // Menyimpan data stok baru
+         // agar setiap user beda beda data
+        Stok::create([
+            'user_id' => Auth::id(),
+            'nama' => $request->nama,
+            'seri' => $request->seri,
+            'ram' => $request->ram,
+            'stok' => $request->stok,
+        ]);
         return redirect()->back()->with('success', 'Stok berhasil ditambahkan.');
     }
 
     public function edit($id)
     {
-        $stok =  Stok::find($id);
+        //agar setiap user beda beda data
+        $stok = Stok::where('user_id', Auth::id())->findOrFail($id);
       
         return view('stock.costom', compact('stok'));
     }
@@ -50,7 +59,10 @@ class StokController extends Controller
             'stok' => 'required|integer',
         ]);
 
-        Stok::where('id', $id)->update([
+        //agar setiap user beda beda data
+
+        $stok = Stok::where('user_id', Auth::id())->findOrFail($id);
+        $stok->update([
             'nama' => $request->nama,
             'seri' => $request->seri,
             'ram' => $request->ram,
@@ -62,7 +74,8 @@ class StokController extends Controller
 
     public function destroy($id)
     {
-        Stok::findOrFail($id)->delete(); // Menghapus stok
+       // Menghapus stok sesuai user berbeda beda
+        Stok::where('user_id', Auth::id())->findOrFail($id)->delete();
         return redirect()->back()->with('success', 'Stok berhasil dihapus.');
     }
 }
